Put SafeEcKeyHandle directly in Interop.EcKey
