+++ src/libraries/Common/src/Interop/Android/System.Security.Cryptography.Native.Android/SafeKeyHandle.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace Microsoft.Win32.SafeHandles
+namespace System.Security.Cryptography
 {
     internal abstract class SafeKeyHandle : SafeHandle
     {
