+++ src/libraries/System.Security.Cryptography.X509Certificates/src/System.Security.Cryptography.X509Certificates.csproj
@@ -419,12 +419,10 @@
              Link="Common\Interop\Android\System.Security.Cryptography.Native.Android\Interop.EcKey.cs" />
     <Compile Include="$(CommonPath)Interop\Android\System.Security.Cryptography.Native.Android\Interop.X509.cs"
              Link="Common\Interop\Android\System.Security.Cryptography.Native.Android\Interop.X509.cs" />
+    <Compile Include="$(CommonPath)Interop\Android\System.Security.Cryptography.Native.Android\SafeKeyHandle.cs"
+             Link="Common\Interop\Android\System.Security.Cryptography.Native.Android\SafeKeyHandle.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\Interop.Libraries.cs"
              Link="Common\Interop\Unix\Interop.Libraries.cs" />
-    <Compile Include="$(CommonPath)Microsoft\Win32\SafeHandles\SafeEcKeyHandle.Android.cs"
-             Link="Common\Microsoft\Win32\SafeHandles\SafeEcKeyHandle.Android.cs" />
-    <Compile Include="$(CommonPath)Microsoft\Win32\SafeHandles\SafeKeyHandle.Android.cs"
-             Link="Common\Microsoft\Win32\SafeHandles\SafeKeyHandle.Android.cs" />
     <Compile Include="$(CommonPath)System\Security\Cryptography\ECAndroid.cs"
              Link="Common\System\Security\Cryptography\ECAndroid.cs" />
     <Compile Include="$(CommonPath)System\Security\Cryptography\ECAndroid.ImportExport.cs"
